feat(ios): refine native list experience
