+++ apps/ios/ZineNative/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,6 +1,15 @@
 {
   "colors": [
     {
+      "color": {
+        "color-space": "srgb",
+        "components": {
+          "alpha": "1.000",
+          "blue": "0.122",
+          "green": "0.400",
+          "red": "0.937"
+        }
+      },
       "idiom": "universal"
     }
   ],
